Added ByteOperationHandler class + other methods
Added the ByteOperationHandler class responsible for handling file reading/writing
operations with the byte/byte[] data structure. Also handles converting
from byte[] to int and merging byte arrays.

Added the TrackerAction class inside the TorrentTrackerClient class. It
contains all the possible actions when it comes to communicating with the
tracker, represented as byte arrays.

Finished the hasConnected method that verifies if the client
has connected in the past and is able to use a cached connection ID
instead of retrieving a new one from the tracker.