Changes to the TorrentTrackerClient class
Finished the connect() method and the hasConnected() method.
Successfully reading the information from cached files and checking
if the transaction id is the same with the initial id.

Started working on the announce() method.
Created the TrackerEvent private class containing all the events
when it comes to the torrent downloading process:

- none
- completed
- started
- stopped